+++ back/src/main/java/com/example/demo/repository/DemandeurRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Demandeur;
 
-public interface DemandeurRepository extends JpaRepository<Demandeur, Long> {
+public interface DemandeurRepository extends JpaRepository<Demandeur, Integer> {
 }
