+++ back/src/main/java/com/example/demo/repository/PassportRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Passport;
 
-public interface PassportRepository extends JpaRepository<Passport, Long> {
+public interface PassportRepository extends JpaRepository<Passport, Integer> {
 }
