+++ back/src/main/java/com/example/demo/repository/PieceJointeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.PieceJointe;
 
-public interface PieceJointeRepository extends JpaRepository<PieceJointe, Long> {
+public interface PieceJointeRepository extends JpaRepository<PieceJointe, Integer> {
 }
