+++ back/src/main/java/com/example/demo/repository/SituationFamilialeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.SituationFamiliale;
 
-public interface SituationFamilialeRepository extends JpaRepository<SituationFamiliale, Long> {
+public interface SituationFamilialeRepository extends JpaRepository<SituationFamiliale, Integer> {
 }
