+++ back/src/main/java/com/example/demo/repository/TypeDemandeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.TypeDemande;
 
-public interface TypeDemandeRepository extends JpaRepository<TypeDemande, Long> {
+public interface TypeDemandeRepository extends JpaRepository<TypeDemande, Integer> {
 }
