+++ back/src/main/java/com/example/demo/repository/TypeVisaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.TypeVisa;
 
-public interface TypeVisaRepository extends JpaRepository<TypeVisa, Long> {
+public interface TypeVisaRepository extends JpaRepository<TypeVisa, Integer> {
 }
