+++ back/src/main/java/com/example/demo/repository/VisaTransformableRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.VisaTransformable;
 
-public interface VisaTransformableRepository extends JpaRepository<VisaTransformable, Long> {
+public interface VisaTransformableRepository extends JpaRepository<VisaTransformable, Integer> {
 }
